Moved TokenTypeRegistry to own .h and .cpp as an attempt to fix the reference error
